requirements.txt
-i https://pypi.python.org/simple
amqp==2.3.2
apipkg==1.5
aniso8601==3.0.2
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
babel==2.6.0
billiard==3.5.0.4
bleach==2.1.4
boto3==1.9.57
botocore==1.12.57
braintree==3.49.0
babel==2.6.0
billiard==3.5.0.4
bleach==2.1.4
boto3==1.9.57
botocore==1.12.57
braintree==3.49.0
celery[redis]==4.2.1
certifi==2018.11.29
chardet==3.0.4
cssselect2==0.2.1
celery[redis]==4.2.1
certifi==2018.11.29
chardet==3.0.4
codecov==2.0.15
coverage==4.5.2
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0 ; python_version >= '3.0'
dj-database-url==0.5.0
dj-email-url==0.1.0
django-appconf==1.0.2
django-babel==0.6.2
django-bootstrap4==0.0.7
django-cc==0.2.2
django-cc-wallet==0.0.1
django-cache-url==3.0.0
django-celery-results==1.0.4
django-countries==5.3.2
django-crispy-forms==1.7.2
django-debug-toolbar-request-history==0.0.9
django-debug-toolbar==1.10.1
django-elasticsearch-dsl==0.5.1
django-filter==2.0.0
django-fsm==2.6.0
django-graphql-jwt==0.2.0
django-extensions==2.1.4
django-graphql-jwt==0.2.0
django-impersonate==1.3
django-js-asset==1.1.0
django-measurement==3.0.0
django-mptt==0.9.1
django-phonenumber-field==2.1.0
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-prices==1.0.4
django-recaptcha==1.4.0
django-redis==4.10.0
django-render-block==0.5
django-silk==2.0.0
django-storages==1.7.1
django-templated-email==2.2.0
django-tables2==2.0.3
django-versatileimagefield==1.10
django-webpack-loader==0.6.0
django-zebra==0.4.5
django==2.1.3
docutils==0.14
elasticsearch-dsl==6.0.1
elasticsearch==6.0.0
execnet==1.5.0
filelock==3.0.10
faker==1.0.0
freezegun==0.3.11
google-i18n-address==2.3.3
google-measurement-protocol==1.0.0
gprof2dot==2016.10.13
graphene-django-optimizer==0.3.5
graphene-django==2.2.0
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
html5lib==1.0.1
idna==2.7
ipaddress==1.0.22
idna==2.7
isort==4.3.4
jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
kombu==4.2.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
multidict==4.5.2
markdown==3.0.1
markupsafe==1.1.0
maxminddb-geolite2==2018.703
maxminddb==1.4.1
measurement==2.0.1
mpmath==1.0.0
oauthlib==2.1.0
phonenumberslite==8.10.1
pillow==5.3.0
prices==1.0.0
promise==2.2.1
purl==1.4
pycodestyle==2.4.0
pycparser==2.19
pygments==2.3.0
pyjwt==1.7.0
pyphen==0.9.5
python-dateutil==2.7.5 ; python_version >= '2.7'
python3-openid==3.1.0 ; python_version >= '3.0'
pytz==2018.7
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
pydocstyle==3.0.0
pylint-celery==0.3
pylint-django==2.0.4
pylint-plugin-utils==0.4
pylint==2.2.2
pytest-cov==2.6.0
pytest-django==3.4.4
pytest-forked==0.2
pytest-mock==1.10.0
pytest-vcr==1.0.1
pytest-xdist==1.24.1
pytest==4.0.1
python-slugify==1.2.6
pyyaml==3.13
requests==2.20.1
raven==6.9.0
razorpay==1.1.1
redis==2.10.6
requests-oauthlib==1.0.0
requests==2.20.1
rx==1.6.1
six==1.11.0
snowballstemmer==1.2.1
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
sqlparse==0.2.4
sympy==1.3
toml==0.10.0
tox==3.5.3
transifex-client==0.13.5
typed-ast==1.1.0 ; python_version < '3.7' and implementation_name == 'cpython'
text-unidecode==1.2
tinycss2==0.6.1
typing==3.6.6
unidecode==1.0.23
urllib3==1.23 ; python_version >= '3.4'
uwsgi==2.0.17.1 ; platform_system != 'Windows'
vine==1.1.4
vcrpy==2.0.1
virtualenv==16.1.0
wrapt==1.10.11
yapf==0.25.0
yarl==1.2.6 ; python_version >= '3.4'
weasyprint==43
webencodings==0.5.1


